Refactor repository interfaces to extend ITenantRepository for better tenant management
- Updated AuditLogRepository, AccessRuleRepository, ChannelRepository, DestinationRepository, and others to inherit from ITenantRepository.
- Removed redundant methods related to tenant-specific operations from these repositories.
- Changed return types for countByTenant methods to size_t for consistency.
- Introduced new use cases for managing clients, document types, schemas, templates, training jobs, and enrichment data.
- Implemented document processing use case with upload and confirmation functionalities.
- Added capabilities retrieval use case for document AI features.

+++ auditlog/source/uim/platform/auditlog/domain/ports/repositories/audit_logs.d
@@ -15,8 +15,6 @@ mixin(ShowModule!());
 /// Port for persisting and querying audit log entries.
 @safe:
 interface AuditLogRepository : ITenantRepository!(AuditLogEntry, AuditLogId) {
-  // AuditLogEntry[] findByTenant(TenantId tenantId);
-// 
   // bool existsById(TenantId tenantId, AuditLogId auditLogId);
   // AuditLogEntry findById(TenantId tenantId, AuditLogId auditLogId);
 
@@ -28,7 +26,7 @@ interface AuditLogRepository : ITenantRepository!(AuditLogEntry, AuditLogId) {
   AuditLogEntry[] search(TenantId tenantId, AuditCategory[] categories,
       long timeFrom, long timeTo, int limit, int offset);
 
-  long countByTenant(TenantId tenantId);
+  size_t countByTenant(TenantId tenantId);
   // void save(TenantId tenantId, AuditLogEntry entry);
   // void save(AuditLogEntry entry);
   // void update(AuditLogEntry entry);
