+++ document-ai/source/uim/platform/document_ai/application/usecases/manage_clients.d
@@ -103,7 +103,7 @@ class ManageClientsUseCase : UIMUseCase {
     return CommandResult(true, id.toString, "");
   }
 
-  long countByTenant(TenantId tenantId) {
+  size_t countByTenant(TenantId tenantId) {
     return repo.countByTenant(tenantId);
   }
 }
