+++ document-ai/source/uim/platform/document_ai/application/usecases/manage_document_types.d
@@ -87,7 +87,7 @@ class ManageDocumentTypesUseCase : UIMUseCase {
     return CommandResult(true, id.toString, "");
   }
 
-  long count(ClientId clientId) {
+  size_t count(ClientId clientId) {
     return repo.countByClient(clientId);
   }
 }
