+++ document-ai/source/uim/platform/document_ai/application/usecases/manage_schemas.d
@@ -121,7 +121,7 @@ class ManageSchemasUseCase : UIMUseCase {
     return CommandResult(true, id.toString, "");
   }
 
-  long count(ClientId clientId) {
+  size_t count(ClientId clientId) {
     return repo.countByClient(clientId);
   }
 }
