+++ document-ai/source/uim/platform/document_ai/application/usecases/manage_templates.d
@@ -119,7 +119,7 @@ class ManageTemplatesUseCase : UIMUseCase {
     return CommandResult(true, id.toString, "");
   }
 
-  long count(ClientId clientId) {
+  size_t count(ClientId clientId) {
     return repo.countByClient(clientId);
   }
 }
